requirements.txt
# requirements.txt is a pip requirements file which describes the necessary
# distributions required for sftrees

# psycopg2, postgis, gdal, geos, proj4...
#python-mapnik #build 0.7.1 from source


Django==1.2.3

-e hg+http://bitbucket.org/ubernostrum/django-registration/#egg=registration
-e hg+http://bitbucket.org/ubernostrum/django-template-utils/#egg=template-utils
xlrd
unidecode
PIL
sorl-thumbnail
geopy
feedparser
python-memcached
BeautifulSoup

#django-debug-toolbar==0.8.3
#django-staticfiles==0.2.0
#django-pagination==1.0.7

simplejson==2.1.1
django-tagging
django-shapes
Tilecache

-e hg+http://bitbucket.org/springmeyer/djmapnik/#egg=djmapnik
